split headers download into own component
-add component statuses
